config/requirements.txt
asgiref==3.8.1
bleach==6.2.0
Django==5.1.6
django-polymorphic==3.1.0
django-taggit==6.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
pillow==11.1.0
PyJWT==2.9.0
python-decouple==3.8
setuptools==75.8.1
sqlparse==0.5.3
webencodings==0.5.1
